# Improve tooltip performance