Tooltips aren't that performant and it's even worse when there are many.
I replaced the linear search in findHoverIndexFromData with binary search, which is faster in my benchmarks:
```
1000 data points
random posX, not necessarily in array
new, 1000000 calls, 125ms
old, 1000000 calls, 1772ms
posX in array
new, 1000000 calls, 103ms
old, 1000000 calls, 1405ms
```

Tooltips are only shown if they are in the browsers viewport.
![tooltip](https://cloud.githubusercontent.com/assets/6831775/22102325/604d59d2-de37-11e6-83f3-df0cab8c702d.gif)

I can't tell if they are actually faster now or if it's just placebo.