@torkelo 2017-01-19
This looks great! 

strange about you noticing it's slow however, what browser are you using? Even with using shared tooltip on multiple graphs (with multiple high resolution series) its butter smooth in chrome for me. 
https://twitter.com/torkelo/status/806095071753990145