@connection-reset 2017-01-19
Firefox 50.1 on Linux, it lags about half a second behind (a bit less now but still noticable).
Of course firefox just isn't as fast as chrome but I'd still like grafana to feel as good in firefox as it does in chrome.