@torkelo 2017-01-19
Yea, have noticed that Grafana performance in Firefox is generally a lot slower than chrome, in some cases by vast margin. Even IE Egde is a lot faster