@connection-reset 2017-01-25
I'd like to look for more performance / general improvements but I've got exams coming up so I can't work on this for a while.