@torkelo 2017-01-25
merged this, but could not notice ANY performance difference in Firefox (on linux) with 10 graphs with 15 series each, but maybe there is noticeable difference on laptops.